Silence various compiler warnings
Many of these are related to the win32 loader, which is barely maintained.
Note that I didn't attempt to fix any bugs.

Fix some format strings that turned out incorrect. (For example, off_t
seems to be 32 bit, and not 64 bit as some code assumes.)

Make configure define __USE_MINGW_ANSI_STDIO on mingw. This is required to
make C99 format specifiers for printf (and similar) work. Add a mingw
patch to mingw.txt. This patch is required to define proper C99 conform
PRI* macros. Since both libav and libass pass format specifiers to
mplayer, it's nearly impossible to fix this with a hack local to mpalyer.

+++ libmpdemux/demux_aac.c
@@ -113,7 +113,7 @@ static int demux_aac_probe(demuxer_t *demuxer)
 	if(cnt < 8)
 		goto fail;
 
-	mp_msg(MSGT_DEMUX, MSGL_V, "demux_aac_probe, INIT: %"PRIu64", PROBED: %"PRIu64", cnt: %d\n", init, probed, cnt);
+	mp_msg(MSGT_DEMUX, MSGL_V, "demux_aac_probe, INIT: %"PRIu64", PROBED: %"PRIu64", cnt: %d\n", (uint64_t)init, (uint64_t)probed, cnt);
 	return DEMUXER_TYPE_AAC;
 
 fail:
